TS: ProteinListFilters/FragmentFilter
